refactor(F3): Change F3 app code from static class to enum

+++ Src/Core/F3/F3.Src/Mapper/F3HttpResponseMapper.cs
@@ -36,7 +36,7 @@ public static class F3HttpResponseMapper
                     new()
                     {
                         HttpCode = StatusCodes.Status200OK,
-                        AppCode = F3Constant.AppCode.SUCCESS,
+                        AppCode = (int)F3Constant.AppCode.SUCCESS,
                     }
             );
